# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Neighborhood Commons API",
    "description": "The Neighborhood Commons API: typed substrate for neighborhood-scale public facts. Five Schema.org-aligned primitives (Place, Organization, Event, Broadcast, List), verification anchored on Type A profile authority, constrained-publishing model with three authority paths (entity-runs-it, pipeline-proxies, witnessed-with-evidence). Read free, write through linked organizations. CC BY 4.0.\n\n**This document is the Spec, part of the Commons Contract.** The Contract is three files together: the Spec (this document — machine-readable, authoritative), the Guide (`/llms.txt` — narrative companion), and the Log (`CHANGELOG.md` — dated record of every change). Rule when they disagree: Spec wins. Guide explains. Log dates.\n\n**3.0.0** establishes the standard four-role event provenance frame (`docs/four-roles.md`) and the type-general provenance method doctrine (`docs/provenance.md`). `source.publisher` is removed (the role is filled by `organizer.name`); `source.method` uses the standard vocabulary (`self_asserted` / `proxied` / `witnessed`); `source.url` carries the proxied source URL when applicable. Organizations, Broadcasts, and Lists gain a `method` field — Organizations admit `seeded` for bulk-imported, awaiting-uptake rows. This is the launch contract; from here forward the substrate is additive-only.",
    "version": "3.0.0",
    "license": {
      "name": "CC-BY-4.0",
...
    },
...
  },
  "servers": [
...
  "paths": {
...
    "/broadcasts": {
      "get": {
        "summary": "List broadcasts",
        "description": "Public read of active Broadcast records (ephemeral signals from organizations, pinned to places). Returns active-only by default; `near`+`radius_km` is the typical proximity filter. Apps surface broadcasts editorially — verification is a filter, not a Commons-side gate.",
        "operationId": "listBroadcasts",
        "parameters": [
          {
            "name": "near",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^-?\\d+\\.?\\d*,-?\\d+\\.?\\d*$"
            },
            "description": "Required for typical proximity reads."
          },
          {
            "name": "radius_km",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0.1,
              "maximum": 100,
              "default": 5
            }
          },
          {
            "name": "organization_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "verified",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter to broadcasts from verified organizations."
          },
          {
            "name": "verified_by",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "not_verified_by",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "created_by_contributor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated broadcast list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "meta",
                    "broadcasts"
                  ],
                  "properties": {
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    },
                    "broadcasts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Broadcast"
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
    "/service/broadcasts": {
      "post": {
        "summary": "Create a broadcast",
        "description": "Create an ephemeral signal pinned to a Place. The Organization must be linked to the calling key (via api_key_organization_links) unless the key has admin privileges. Verification is NOT required at create time — apps filter on verified status when surfacing.",
        "operationId": "serviceCreateBroadcast",
        "security": [
          {
            "serviceApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Broadcast created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "broadcast"
                  ],
                  "properties": {
                    "broadcast": {
                      "$ref": "#/components/schemas/Broadcast"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Not linked to this organization"
          }
        }
      }
    },
    "/service/broadcasts/{id}/retract": {
      "post": {
        "summary": "Retract an active broadcast",
        "description": "Marks the broadcast retracted. Status changes to `retracted` and `retracted_at` is set. Idempotent.",
        "operationId": "serviceRetractBroadcast",
        "security": [
          {
            "serviceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retracted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "broadcast"
                  ],
                  "properties": {
                    "broadcast": {
                      "$ref": "#/components/schemas/Broadcast"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Not linked to this organization"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Meta": {
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "spec",
          "license"
        ],
        "properties": {
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "spec": {
            "type": "string",
            "example": "neighborhood-api-v0.2",
            "description": "The upstream Neighborhood API spec this implementation conforms to (distinct from the implementation version of this specific Commons instance, which is in `info.version` of `/openapi.json` and at `/meta` under `implementation_version`)."
          },
          "license": {
            "type": "string",
            "example": "CC-BY-4.0"
          }
        }
      },
...
      "Source": {
        "type": "object",
        "description": "Event provenance under the four-role frame (see docs/four-roles.md). Carries the contributor identity (which ecosystem participant routed the event in), the method (authority shape: self_asserted, proxied, witnessed), the source URL when applicable (proxied events), plus collected_at and license. Does not carry a `publisher` field — the role 'who is this from?' is filled by the top-level `organizer.name`.",
        "required": [
          "method",
          "url",
          "contributor",
          "collected_at",
          "license"
        ],
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "self_asserted",
              "proxied",
              "witnessed"
            ],
            "description": "Standard provenance method (docs/provenance.md). `self_asserted`: the organizer asserted this via the contributor (entity-runs-it). `proxied`: the contributor extracted this from a public URL (pipeline-proxies). `witnessed`: the contributor observed this with documentary evidence under a collective identity (witnessed-with-evidence). Drives rendering rules — consumers typically suppress the 'via {contributor}' line on `witnessed` events because the organizer is the collective constituted by the contributor."
          },
          "url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri",
            "description": "When method is `proxied`, the public URL the contributor extracted from (for transparency). Null otherwise."
          },
          "contributor": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Public-facing name of the contributing app/pipeline."
                  },
                  "url": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uri",
                    "description": "Contributor website."
                  }
                }
              }
            ],
            "description": "The ecosystem participant that routed this event into the Commons. Editorial public-facing identity — separate from the operational API key. Null for legacy rows without contributor attribution."
          },
          "collected_at": {
            "type": "string",
            "format": "date-time"
          },
          "license": {
            "type": "string",
            "example": "CC BY 4.0"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "$ref": "#/components/schemas/ErrorCode"
              },
              "message": {
                "type": "string"
              },
              "domain": {
                "type": "string",
                "description": "Present when code is DOMAIN_PENDING_REVIEW"
              }
            }
          }
        }
      },
      "PostalAddress": {
        "type": "object",
        "description": "Schema.org PostalAddress. Used to represent structured addresses on Place and related types.",
        "properties": {
          "streetAddress": {
            "type": [
              "string",
              "null"
            ]
          },
          "addressLocality": {
            "type": [
              "string",
              "null"
            ],
            "description": "City / locality, e.g. \"Philadelphia\""
          },
          "addressRegion": {
            "type": [
              "string",
              "null"
            ],
            "description": "State or region. US: 2-letter code (e.g. \"PA\")"
          },
          "postalCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "addressCountry": {
            "type": "string",
            "description": "ISO 3166-1 alpha-2 code",
            "default": "US"
          }
        }
      },
      "GeoCoordinates": {
        "type": "object",
        "description": "Schema.org GeoCoordinates.",
        "required": [
          "latitude",
          "longitude"
        ],
        "properties": {
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          }
        }
      },
      "IdentifierValue": {
        "type": "object",
        "description": "Schema.org PropertyValue used in identifier arrays. Carries external IDs (e.g., googlePlaceId) alongside our internal UUIDs.",
        "required": [
          "propertyID",
          "value"
        ],
        "properties": {
          "propertyID": {
            "type": "string",
            "example": "googlePlaceId"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "OpeningHoursEntry": {
        "type": "object",
        "description": "Schema.org OpeningHoursSpecification entry. One per day-range. Times are 24h HH:MM in the entity's local timezone.",
        "properties": {
          "dayOfWeek": {
            "type": [
              "string",
              "array"
            ],
            "description": "Single day or array of days. Schema.org-canonical: full names.",
            "items": {
              "type": "string",
              "enum": [
                "Monday",
                "Tuesday",
                "Wednesday",
                "Thursday",
                "Friday",
                "Saturday",
                "Sunday"
              ]
            }
          },
          "opens": {
            "type": "string",
            "description": "24h time, e.g. \"09:00\""
          },
          "closes": {
            "type": "string",
            "description": "24h time, e.g. \"17:00\""
          }
        }
      },
      "Place": {
        "type": "object",
        "description": "Schema.org Place. A physical location. Identity is the address; the stable cross-source key is `googlePlaceId` (the one Google datum permitted indefinite storage). Categories come from OpenStreetMap (ODbL-licensed, attributed) — see `placeCategories` and `categorySource`. Google response data beyond `googlePlaceId` is never persisted into these columns.",
        "required": [
          "id",
          "name",
          "geo"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PostalAddress"
              },
              {
                "type": "null"
              }
            ]
          },
          "geo": {
            "$ref": "#/components/schemas/GeoCoordinates"
          },
          "identifier": {
            "type": "array",
            "description": "External IDs. Common entry: { propertyID: \"googlePlaceId\", value: \"ChIJ...\" }",
            "items": {
              "$ref": "#/components/schemas/IdentifierValue"
            }
          },
          "placeCategories": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Place categorization (e.g., \"cafe\", \"live_music_venue\"). Sourced from OpenStreetMap by default; never from Google response data."
          },
          "categorySource": {
            "type": ["string", "null"],
            "enum": ["osm", "admin_review", "publisher_declaration", null],
            "description": "Provenance of `placeCategories`. `osm` (default), `admin_review` (operator added), `publisher_declaration` (verified publisher refined)."
          },
          "region_slug": {
            "type": [
              "string",
              "null"
            ],
            "description": "Slug of the region this place falls within, if any"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Verification": {
        "type": "object",
        "description": "Verification block exposed on verified Organization responses. Anchors Type A authority on organizations only — there is no cross-app reputation graph.",
        "required": [
          "method",
          "verifiedAt",
          "verifiedByApp"
        ],
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "domain_email_loop",
              "manual_review"
            ]
          },
          "verifiedVia": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "in_person",
              "video_call",
              null
            ],
            "description": "Present for manual_review only."
          },
          "verifiedAt": {
            "type": "string",
            "format": "date-time"
          },
          "verifiedByApp": {
            "type": "string",
            "description": "Stable app name (snapshot at approval time, immune to key rotation)."
          }
        }
      },
...
      "Organization": {
        "type": "object",
        "description": "Schema.org Organization. The unified entity primitive — businesses, community groups, nonprofits, collectives, solo operators (organizations-of-one). Classification emerges from `tags` (descriptive labels), `commercial` (for-profit boolean), and structural signals (primary place, event history). The legacy `kind` enum was retired because it mixed structural facts, vibes, and legal status into one false choice. `additionalType` is derived structurally — `LocalBusiness` when a primary place is set, plain `Organization` otherwise. `method` carries the provenance under the standard four-value vocabulary (docs/provenance.md).",
        "required": [
          "id",
          "slug",
          "name",
          "tags",
          "method"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "method": {
            "type": "string",
            "enum": [
              "self_asserted",
              "proxied",
              "witnessed",
              "seeded"
            ],
            "description": "Standard provenance method (docs/provenance.md). `self_asserted`: first-party authority (the org has claimed and verified itself). `proxied`: extracted from a public source (e.g. a business registry). `witnessed`: collective-evidence path. `seeded`: bulk-imported, awaiting first-party uptake. Consumers can filter for `self_asserted` orgs to surface only first-party-asserted records."
          },
          "legalName": {
            "type": [
              "string",
              "null"
            ],
            "description": "Official registered name. Optional, primarily for verified businesses."
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 50
            },
            "description": "Descriptive labels — free-form within format rules. Recommended starter vocabulary at `/v1/meta/tags` (when shipped). Not a hard taxonomy; consumer apps filter on whatever tags appear in practice."
          },
          "commercial": {
            "type": ["boolean", "null"],
            "description": "For-profit (true) vs. non-profit/community (false). Null = unspecified. Replaces the structural-axis component of the legacy `kind` enum."
          },
          "additionalType": {
            "type": "string",
            "format": "uri",
            "description": "Schema.org type URL derived from structural data: `https://schema.org/LocalBusiness` when the organization has a primary place, `https://schema.org/Organization` otherwise. Apps wanting richer subtyping derive from `tags` or `place_categories`."
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "logo": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "image": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "telephone": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "format": "email"
          },
          "sameAs": {
            "type": "array",
            "description": "URLs that unambiguously identify this organization (Wikipedia, Wikidata, social profiles).",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "openingHoursSpecification": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/OpeningHoursEntry"
            }
          },
          "location": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Place"
              },
              {
                "type": "null"
              }
            ],
            "description": "The Place this Organization primarily operates at. Null for organizations without a fixed location (touring, online-only)."
          },
          "verified": {
            "type": "boolean"
          },
          "verification": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Verification"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Broadcast": {
        "type": "object",
        "description": "Ephemeral real-time signal from an Organization, pinned to a Place. Maximum lifetime 24h. No direct Schema.org analog; conventions borrowed from SpecialAnnouncement (`datePosted`, `expires`). `method` carries the standard provenance vocabulary — broadcasts are always first-party from the organization (only `self_asserted` is valid today).",
        "required": [
          "id",
          "message",
          "datePosted",
          "expires",
          "organization",
          "location",
          "source",
          "method"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 280
          },
          "datePosted": {
            "type": "string",
            "format": "date-time"
          },
          "expires": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "expired",
              "retracted"
            ]
          },
          "method": {
            "type": "string",
            "enum": [
              "self_asserted"
            ],
            "description": "Standard provenance method (docs/provenance.md). Only `self_asserted` is valid today — broadcasts are always first-party from the organization. Field exists for symmetry across primitives; new methods can be added additively."
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "location": {
            "$ref": "#/components/schemas/Place"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          }
        }
      },
      "BroadcastInput": {
        "type": "object",
        "required": [
          "organizationId",
          "placeId",
          "message",
          "expires"
        ],
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "placeId": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 280
          },
          "expires": {
            "type": "string",
            "format": "date-time",
            "description": "Must be in the future and within 24h of now."
          }
        }
      },
...
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Request payload or query parameters failed validation",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "browseApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Optional for read endpoints (higher rate limit bucket). Required for webhook management."
      },
      "serviceApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Service-tier API key. Scoped: can only modify data for accounts linked to this key. Admin keys bypass scoping."
      }
    }
  }
}